Updated VoIPApplication
BUFFER_SIZE constant has been changed to a non-final static field and
subsequently renamed to bufferSize. Added several instance fields
including a GUI instance and Scanner instance (scanner will be removed
in the future as the GUI improves). A shutdown() method has been added.
This is mostly a wrapper to endCall(), another new method, but it also
will call System.exit.  The endCall() method will call
client.stopClient() and server.stopServer() in attempt to gracefully end
the call with the other peer. The new getCallInfo() method creates a
string showing the client's address and elapsed time of the call. This
string is displayed on the GUI's main JLabel when called by the
UpdateLabelTask. Last to be noted is the isInCall() method which returns
true if the peerIp is not null, the client is running, and the server is
running.